chore: separated IS analysis from figures
